Agent Frequency can now be < 1Hz
Changed int value of frequency to fload

+++ src/architecture/agentConfig.h
@@ -9,7 +9,7 @@ class AgentConfig {
 
         int Priority;
 
-        int ThreadFrequency; 
+        float ThreadFrequency;
 
         int StatsCycle;
 };
